Defaults for declared but unset env vars.
Make sure defaults are applied for declared-but-empty environment
variables.

This is an issue because adding them to LAMBDA_ENV_VARS (as is done by
default in .env.lambda) causes AWS to set unset variables to None.  Our
calls to os.getenv('foo', 'bar') would then not result in 'bar' but None
instead.

Specifically, KODI_SCHEME and KODI_SUBPATH are unset by default in
.env.lambda, which would result in a URL like this:

  http://None:/xxx.xxxx.xxx:8080/None/jsonrpc

See here for the initial report of this bug:

  http://forum.kodi.tv/showthread.php?tid=254502&pid=2449860#pid2449860