Segment Playlist adds.
Kodi has a limit on how much it will accept in a single playlist add
call, so make multiple calls if necessary.

NOTE that this is incredibly slow, but because there is no other way to
shuffle playlists from Kodi's JSON-RPC interface, this is all we can do
at the moment.

Addresses Issue #79.