Search top matches for songs and albums.
If told to search for x on y, try to find x for all of the matches of y
instead of just the first one.

The idea is to allow for songs to be found on duplicate albums and for
songs and albums to be found by similar/duplicate artists.